requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
mysqlclient==1.4.6
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.0.4